part4/hbnb/requirements.txt
flask
flask-restx
flask-cors
PyJWT
